+++ gmtsar/csh/prep_data_linux.csh
@@ -9,6 +9,8 @@
 #
 # set a local directory that stores S1A and S1B orbits. e.g. orb_dir/S1A and orb_dir/S1B
 #
+source ~/.cshrc
+
 set orb_dir = "/geosat2/InSAR_Processing/Sentinel_Orbits"
 
 rm data.in
@@ -36,16 +38,18 @@ foreach line ( ` awk '{ print $0 }' < text.dat ` )
     # uses date for time manipulation -ben
     #set n1 = `date -v-1d -jf "%Y%m%d" $mstem +%Y%m%d`
     #set n2 = `date -v+1d -jf "%Y%m%d" $mstem +%Y%m%d`
-    set n1 = ` date --date="$date0 - 1 day" +%Y%m%d `
-    set n2 = ` date --date="$date0 + 1 day" +%Y%m%d `
+    set n1 = ` date --date="$mstem - 1 day" +%Y%m%d `
+    set n2 = ` date --date="$mstem + 1 day" +%Y%m%d `
     set SAT = `echo $mname | awk '{print toupper(substr($1,1,3))}'`
 
     #cp ../../../../orbit/*$n1*$n2* .
     set orb = `grep $SAT orbits.list | grep $n1 | grep $n2`
-    if (-f $orb_dir/$SAT/$orb) then 
-      cp $orb_dir/$SAT/$orb .
-    else
-      wget $url_root/$orb
+    if (! -f $orb) then
+      if (-f $orb_dir/$SAT/$orb) then 
+        cp $orb_dir/$SAT/$orb .
+      else
+        wgetasf $url_root/$orb
+      endif
     endif
     #if (! -f $orb) wget $url_root"/"$orb
     #set orb = `ls *$n1*$n2*`
@@ -67,16 +71,18 @@ foreach line ( ` awk '{ print $0 }' < text.dat ` )
 end
 #set n1 = `date -v-1d -jf "%Y%m%d" $mstem +%Y%m%d`
 #set n2 = `date -v+1d -jf "%Y%m%d" $mstem +%Y%m%d`
-set n1 = ` date --date="$date0 - 1 day" +%Y%m%d `
-set n2 = ` date --date="$date0 + 1 day" +%Y%m%d `
+set n1 = ` date --date="$mstem - 1 day" +%Y%m%d `
+set n2 = ` date --date="$mstem + 1 day" +%Y%m%d `
 set SAT = `echo $mname | awk '{print toupper(substr($1,1,3))}'`
 echo "Writing record $mstem"
 #set orb = `ls *$n1*$n2*`
 set orb = `grep $SAT orbits.list | grep $n1 | grep $n2`
-if (-f $orb_dir/$SAT/$orb) then
-  cp $orb_dir/$SAT/$orb .
-else
-  wgetasf $url_root/$orb
+if (! -f $orb) then 
+  if (-f $orb_dir/$SAT/$orb) then
+    cp $orb_dir/$SAT/$orb .
+  else
+    wgetasf $url_root/$orb
+  endif
 endif
 #if (! -f $orb)  wget $url_root"/"$orb
 #set orb = `ls *$mstem*EOF`
